Improvements and refactoring.
